requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.13.0
pymupdf==1.23.8
pdfplumber==0.10.3
pandas==2.1.3
openpyxl==3.11.0
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
Pillow==10.1.0
